extracted examples
