+++ typescript/examples/langchain/structured-chat-agent.ts
@@ -1,10 +1,10 @@
-import HederaAgentLangchainToolkit from '@/langchain/toolkit.js';
+import { HederaLangchainToolkit, AgentMode } from 'hkav3';
 import { ChatOpenAI } from '@langchain/openai';
-import { ChatPromptTemplate } from '@langchain/core/prompts';
-import { AgentExecutor, createToolCallingAgent } from 'langchain/agents';
+import type { ChatPromptTemplate } from '@langchain/core/prompts';
+import { pull } from 'langchain/hub';
+import { AgentExecutor, createStructuredChatAgent } from 'langchain/agents';
 import { BufferMemory } from 'langchain/memory';
 import { Client, PrivateKey } from '@hashgraph/sdk';
-import { AgentMode } from '@/shared/configuration.js';
 import prompts from 'prompts';
 import * as dotenv from 'dotenv';
 dotenv.config();
@@ -22,7 +22,7 @@ async function bootstrap(): Promise<void> {
   );
 
   // Prepare Hedera toolkit (load all tools by default)
-  const hederaAgentToolkit = new HederaAgentLangchainToolkit({
+  const hederaAgentToolkit = new HederaLangchainToolkit({
     client,
     configuration: {
       tools: [], // load every available tool
@@ -33,22 +33,16 @@ async function bootstrap(): Promise<void> {
   });
 
   // Load the structured chat prompt template
-  const prompt = ChatPromptTemplate.fromMessages([
-    ['system', 'You are a helpful assistant'],
-    ['placeholder', '{chat_history}'],
-    ['human', '{input}'],
-    ['placeholder', '{agent_scratchpad}'],
-  ]);
+  const prompt = await pull<ChatPromptTemplate>('hwchase17/structured-chat-agent');
 
   // Fetch tools from toolkit
   const tools = hederaAgentToolkit.getTools();
 
-  // Create the underlying agent
-  const agent = createToolCallingAgent({
+  const agent = await createStructuredChatAgent({
     llm,
     tools,
     prompt,
-  });
+  } as any);
 
   // In-memory conversation history
   const memory = new BufferMemory({
@@ -64,7 +58,7 @@ async function bootstrap(): Promise<void> {
     tools,
     memory,
     returnIntermediateSteps: false,
-  });
+  } as any);
 
   console.log('Hedera Agent CLI Chatbot — type "exit" to quit');
