+++ typescript/examples/langchain/tsconfig.json
@@ -1,0 +1,18 @@
+{
+    "compilerOptions": {
+        "module": "CommonJS",
+        "moduleResolution": "node",
+        "target": "ES2020",
+        "esModuleInterop": true,
+        "allowSyntheticDefaultImports": true,
+        "skipLibCheck": true,
+        "skipDefaultLibCheck": true,
+        "disableSizeLimit": true,
+        "noErrorTruncation": true,
+        "maxNodeModuleJsDepth": 2,
+        "types": [
+            "node"
+        ],
+        "strict": true,
+    }
+}
